requirements.txt
streamlit>=1.31.0
opencv-python>=4.9.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
jsonlines>=4.0.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
openai>=1.0.0
scenedetect[opencv]>=0.6.0
python-dotenv>=1.0.0

# M2 dependencies - Multi-modal retrieval
open-clip-torch>=2.24.0
torch>=2.1.0
torchvision>=0.16.0
faiss-cpu>=1.7.4
tabulate>=0.9.0
